Fix issue with number of ROIs when using atlas.
When checking the number of rois by calling int(atlasdata.max()),
sometimes the count would be off by one. I didn't figure out
what the problem was, but it could have something to do with pixdim[0]
of some atlases because nibabel was throwing warnings.
By querying the number of rois from the shape of the signal, we
guarantee that rsns and mixing will have the same size.